:sparkles: feat: On delete employee check product reference

+++ src/employee/services/employee.service.ts
@@ -3,10 +3,12 @@ import { UpdateEmployeeDTO } from "../../dtos/updateEmployee.dto";
 import { CreateEmployeeDTO } from "../../dtos/createEmployee.dto";
 import { Employee } from "../entities/employee.entity";
 import { EmployeeRepository } from "../repositories/employee.repository";
+import { ProductRepository } from "src/product/repositories/product.repository";
 
 @Injectable()
 export class EmployeeService {
-  constructor(private readonly employeeRepository: EmployeeRepository) {}
+  constructor(private readonly employeeRepository: EmployeeRepository, 
+    private readonly productRepository: ProductRepository) {}
 
   async findAll(): Promise<Employee[]> {
     const findedEmployees = await this.employeeRepository.findAll();
@@ -37,10 +39,17 @@ export class EmployeeService {
   }
 
   async delete(uuid: string): Promise<void> {
-    const findedUser = await this.employeeRepository.findOne(uuid);;
+    const findedUser = await this.employeeRepository.findOne(uuid);
+    const isReferenced = await this.productRepository.findByEmployee(uuid);
+
     if (!findedUser) {
       throw new NotFoundException(`Employee with uuid ${uuid} not found`);
     }
+
+    if (isReferenced) {
+      throw new BadRequestException(`Employee with uuid ${uuid} is referenced in one or more products`);
+    }
+
     return this.employeeRepository.delete(uuid);
   }
 }
